需求[#18191]: 任务管理界面 任务下发 周期与定时任务下发给摄像头
    [描述说明]:任务管理界面 任务下发 周期与定时任务下发给摄像头
    [问题原因]:
    [解决方法]:
    [测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/device/service/TCruisePointInstanceService.java
@@ -295,7 +295,10 @@ public class TCruisePointInstanceService{
         //巡检点配置
         if (tCruisePointInstanceDetail.getIds().size()!=0){
             for (Long id:tCruisePointInstanceDetail.getIds()) {
-                haveList.remove(id);
+                if(haveList != null && haveList.size()>0){
+                    haveList.remove(id);
+                }
+                
                 TCruisePointAttr tCruisePointAttr = new TCruisePointAttr();
                 if(cruiseType == 229){//视频
                     TCameraPreset tCameraPreset = tCameraPresetDao.selectByPrimaryId(id);
@@ -326,8 +329,10 @@ public class TCruisePointInstanceService{
             }
         }
         //删除关联表的巡检实例
-        tCruisePlanAttrDao.deleteByInstanceId(haveList);
-        tCruiseTaskAttrDao.deleteByInstanceId(haveList);
+        if(haveList != null && haveList.size()>0){
+            tCruisePlanAttrDao.deleteByInstanceId(haveList);
+            tCruiseTaskAttrDao.deleteByInstanceId(haveList);
+        }
         return result;
     }
